Add files via upload
This is a build on data.csv. It includes additional columns for the lag (recency) in days since the last successful day of completing exercises, frequency in days in the past on which exercise was done successfully. Also columns for normlag which is lag/25 (the max no. of lag days) and normfre which is freq/11 (max frequency).